galeeza/pyproject.toml
[tool.poetry]
name = "galeeza"
version = "0.1.0"
description = ""
authors = ["Dani Bachini <danielle.bachini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
bcrypt = "^4.0.1"
mysql-connector-python = "^8.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
